# hub/public/swagger.json
{
    "components": {
...
        "schemas": {
            "LabDto": {
                "properties": {
                    "domainId": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "country": {
                        "type": "string"
                    },
                    "components": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                }
            },
            "NoIdLabDto": {
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "country": {
                        "type": "string"
                    },
                    "components": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                }
            },
...
        },
        "securitySchemes": {}
    },
    "info": {
        "title": "DICYSTECH API",
        "version": "0.1",
        "contact": {}
    },
    "openapi": "3.0.0",
    "paths": {
...
        "/api/lab": {
            "post": {
                "operationId": "AddLab",
                "responses": {
                    "201": {
                        "description": "Added Lab",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LabDto"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request: Validation failed"
                    }
                },
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/NoIdLabDto"
                            }
                        }
                    }
                },
                "tags": ["Lab"]
            },
            "get": {
                "operationId": "GetAllLabs",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/LabDto"
                                    }
                                }
                            }
                        }
                    }
                },
                "tags": ["Lab"]
            }
        },
        "/api/lab/{name}": {
            "get": {
                "operationId": "GetLabByName",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LabDto"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Lab not found"
                    }
                },
                "tags": ["Lab"]
            }
        },
...
    }
}